refactor(direct-control): extract narrative choice source
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only the App UI narrative choice command builder and embedded source into packages/civ7-direct-control/src/play/operations/narrative-request.ts while keeping the public facade export in index.ts. The facade injects approval, App UI execution, notification reads, player-operation validation, payload parsing, ComponentID validation, and serialization.

Preserves CHOOSE_NARRATIVE_STORY_DIRECTION send behavior, narrative popup/panel closeout calls, focused narrative package/CLI proof, and existing postcondition classification. Runtime/live-game proof, telemetry, AI ingestion, semantic CLI projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- narrative-choice.test.ts
- bun run --cwd packages/cli test -- game.play.narrative.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership and no-cycle rg scans

Local package/source relocation proof only; no runtime/live-game, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
