+++ packages/civ7-direct-control/src/index.ts
@@ -3433,11 +3433,11 @@ export async function requestCiv7NarrativeChoice(
     validatePlayerId,
     assertComponentId: assertCiv7ComponentId,
     executeAppUiCommand: executeCiv7AppUiCommand,
-    buildNarrativeChoiceRequestCommand,
     getPlayNotificationView: getCiv7PlayNotificationView,
     canStartPlayerOperation: canStartCiv7PlayerOperation,
     parseNarrativePayload: (result, label) =>
       jsonPayloadFromCommandResult<Civ7NarrativeChoiceCommandPayload>(result, label),
+    jsLiteral,
     invalidTargetTypeError: () => {
       throw new Civ7DirectControlError("command-failed", "targetType is required");
     },
@@ -3971,13 +3971,6 @@ function buildCultureChoiceCloseoutCommand(input: Civ7CultureChoiceCloseoutInput
   })()`;
 }
 
-function buildNarrativeChoiceRequestCommand(input: Civ7NarrativeChoiceInput): string {
-  return `(() => {
-    ${narrativeChoiceRequestSource()}
-    return JSON.stringify(sendNarrativeChoice(${jsLiteral(input)}));
-  })()`;
-}
-
 function buildNotificationDismissalCommand(input: Civ7NotificationDismissInput, options: { send: boolean; verificationAttempts?: number }): string {
   return `(() => {
     ${notificationDismissalSource()}
@@ -3993,127 +3986,6 @@ function probeHelperSource(): string {
       } catch (err) {
         return { ok: false, error: String(err) };
       }
-    };`;
-}
-
-function narrativeChoiceRequestSource(): string {
-  return `${probeHelperSource()}
-    const toComponentId = (value) => {
-      if (!value || typeof value !== "object") return null;
-      if (typeof value.owner !== "number" || typeof value.id !== "number") return null;
-      const out = { owner: value.owner, id: value.id };
-      if (typeof value.type === "number") out.type = value.type;
-      return out;
-    };
-    const sameComponentId = (left, right) => {
-      const a = toComponentId(left);
-      const b = toComponentId(right);
-      if (!a || !b) return false;
-      return a.owner === b.owner && a.id === b.id && (a.type ?? null) === (b.type ?? null);
-    };
-    const safeCall = (label, fn) => {
-      try {
-        return { ok: true, value: fn() };
-      } catch (err) {
-        return { ok: false, error: label + ": " + String(err) };
-      }
-    };
-    const narrativeSelectors = [
-      "small-narrative-event",
-      "graphic-narrative-event",
-      "screen-narrative-event",
-      "screen-narrative-trial",
-    ];
-    const viewMethod = () => typeof UIViewChangeMethod !== "undefined" ? UIViewChangeMethod.PlayerInteraction : undefined;
-    const narrativePopupManager = () => typeof NarrativePopupManager === "undefined" ? null : NarrativePopupManager;
-    const narrativePanels = () => {
-      if (typeof document === "undefined") return [];
-      return narrativeSelectors.flatMap((selector) => Array.from(document.querySelectorAll?.(selector) ?? []));
-    };
-    const summarizePanel = (root) => {
-      const component = root?._component ?? null;
-      return {
-        panelType: root?.tagName ?? null,
-        componentType: component?.constructor?.name ?? null,
-        targetStoryId: toComponentId(component?.targetStoryId),
-        storyType: component?.storyType ?? null,
-        choiceKeys: Array.from(root?.querySelectorAll?.("fxs-reward-button[small-narrative-choice-key]") ?? [])
-          .map((button) => button.getAttribute("small-narrative-choice-key"))
-          .filter(Boolean),
-      };
-    };
-    const readNarrativeUiState = (target) => {
-      const panels = narrativePanels().map(summarizePanel);
-      const matchingPanels = panels.filter((panel) => sameComponentId(panel.targetStoryId, target));
-      return {
-        panelCount: panels.length,
-        panels,
-        matchingPanelCount: matchingPanels.length,
-        matchingPanels,
-        popupShowing: safeCall("NarrativePopupManager.isShowing", () => narrativePopupManager()?.isShowing?.()),
-        currentNarrativeData: safeCall("NarrativePopupManager.currentNarrativeData", () => narrativePopupManager()?.currentNarrativeData ?? null),
-      };
-    };
-    const closeVisibleNarrativePanels = (target) => safeCall("visible narrative panel close", () => {
-      const panels = narrativePanels()
-        .filter((root) => sameComponentId(root?._component?.targetStoryId, target));
-      const method = viewMethod();
-      const results = panels.map((root) => {
-        const component = root?._component ?? null;
-        if (typeof component?.close !== "function") {
-          return { panelType: root?.tagName ?? null, closed: false, reason: "panel component has no close function" };
-        }
-        component.close(method);
-        return { panelType: root?.tagName ?? null, closed: true };
-      });
-      return { attempted: panels.length, results };
-    });
-    const closeNarrativePopup = () => safeCall("NarrativePopupManager.closePopup", () => {
-      const manager = narrativePopupManager();
-      if (!manager || typeof manager.closePopup !== "function") return { available: false };
-      manager.closePopup();
-      return { available: true };
-    });
-    const sendNarrativeChoice = (input) => {
-      const localPlayerId = GameContext.localPlayerID;
-      const playerId = localPlayerId;
-      const args = { TargetType: input.targetType, Target: input.target, Action: input.action };
-      const before = readNarrativeUiState(input.target);
-      const canStart = safeCall("Game.PlayerOperations.canStart", () => Game.PlayerOperations.canStart(
-        playerId,
-        PlayerOperationTypes.CHOOSE_NARRATIVE_STORY_DIRECTION,
-        args,
-        false,
-      ));
-      let sent = false;
-      let sendResult = null;
-      if (canStart.ok && (canStart.value?.Success === true || canStart.value?.canStart === true)) {
-        sendResult = safeCall("Game.PlayerOperations.sendRequest", () => Game.PlayerOperations.sendRequest(
-          playerId,
-          PlayerOperationTypes.CHOOSE_NARRATIVE_STORY_DIRECTION,
-          args,
-        ));
-        sent = sendResult.ok === true;
-      }
-      const popupClose = sent ? closeNarrativePopup() : { ok: false, skipped: true, reason: "operation was not sent" };
-      const panelClose = sent ? closeVisibleNarrativePanels(input.target) : { ok: false, skipped: true, reason: "operation was not sent" };
-      return {
-        localPlayerId,
-        playerId,
-        args,
-        canStart,
-        sent,
-        sendResult,
-        ui: {
-          before,
-          after: readNarrativeUiState(input.target),
-          panelClose,
-          popupClose,
-        },
-        notes: [
-          "This mirrors the official narrative button handler: CHOOSE_NARRATIVE_STORY_DIRECTION, NarrativePopupManager.closePopup, and visible narrative panel close.",
-        ],
-      };
     };`;
 }
